Split contract resolvers from json utility
