+++ src/docfx/lib/json/JsonUtility.cs
@@ -2,12 +2,9 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.IO;
 using System.Linq;
-using System.Reflection;
 using System.Text.RegularExpressions;
 using System.Threading;
 using Newtonsoft.Json;
@@ -37,7 +34,7 @@ namespace Microsoft.Docs.Build
         {
             NullValueHandling = NullValueHandling.Ignore,
             Converters = s_jsonConverters,
-            ContractResolver = new SchemaValidationContractResolver { NamingStrategy = s_namingStrategy },
+            ContractResolver = new SchemaContractResolver { NamingStrategy = s_namingStrategy },
         });
 
         private static readonly JsonSerializer s_indentSerializer = JsonSerializer.Create(new JsonSerializerSettings
@@ -50,18 +47,12 @@ namespace Microsoft.Docs.Build
 
         private static readonly ThreadLocal<Stack<Status>> t_status = new ThreadLocal<Stack<Status>>(() => new Stack<Status>());
 
-        // HACK: Json.NET property deserialization is case insensitive:
-        // https://github.com/JamesNK/Newtonsoft.Json/issues/815,
-        // Force property deserialization to be case sensitive by hijacking GetClosestMatchProperty implementation.
-        private static readonly Action<JsonPropertyCollection, List<JsonProperty>> s_makeJsonCaseSensitive =
-            ReflectionUtility.CreateInstanceFieldSetter<JsonPropertyCollection, List<JsonProperty>>("_list");
-
-        private static readonly List<JsonProperty> s_emptyPropertyList = new List<JsonProperty>();
-
         private static readonly Action<JToken, int, int> s_setLineInfo =
             ReflectionUtility.CreateInstanceMethod<JToken, Action<JToken, int, int>>("SetLineInfo", new[] { typeof(int), typeof(int) });
 
         internal static JsonSerializer Serializer => s_serializer;
+
+        internal static Status State => t_status.Value.Peek();
 
         static JsonUtility()
         {
@@ -519,129 +510,7 @@ namespace Microsoft.Docs.Build
             return null;
         }
 
-        private static void MakePropertyCollectionCaseSensitive(JsonPropertyCollection properties)
-        {
-            s_makeJsonCaseSensitive(properties, s_emptyPropertyList);
-        }
-
-        private class JsonContractResolver : DefaultContractResolver
-        {
-            protected override JsonObjectContract CreateObjectContract(Type objectType)
-            {
-                var contract = base.CreateObjectContract(objectType);
-                MakePropertyCollectionCaseSensitive(contract.Properties);
-                return contract;
-            }
-
-            protected override JsonProperty CreateProperty(MemberInfo member, MemberSerialization memberSerialization)
-            {
-                var prop = base.CreateProperty(member, memberSerialization);
-                ShouldNotSerializeEmptyArray();
-                SetFieldWritable();
-                return prop;
-
-                void ShouldNotSerializeEmptyArray()
-                {
-                    if (typeof(IEnumerable).IsAssignableFrom(prop.PropertyType) && !(prop.PropertyType == typeof(string)))
-                    {
-                        prop.ShouldSerialize =
-                        target =>
-                        {
-                            var value = prop.ValueProvider.GetValue(target);
-
-                            if (value is IEnumerable enumer && !enumer.GetEnumerator().MoveNext())
-                            {
-                                return false;
-                            }
-
-                            return true;
-                        };
-                    }
-                }
-
-                void SetFieldWritable()
-                {
-                    if (!prop.Writable)
-                    {
-                        if (member is FieldInfo f && f.IsPublic && !f.IsStatic)
-                        {
-                            prop.Writable = true;
-                        }
-                    }
-                }
-            }
-        }
-
-        private sealed class SchemaValidationContractResolver : JsonContractResolver
-        {
-            protected override JsonProperty CreateProperty(MemberInfo member, MemberSerialization memberSerialization)
-            {
-                var prop = base.CreateProperty(member, memberSerialization);
-                var converter = GetConverter(member);
-
-                if (converter != null)
-                {
-                    prop.Converter = converter;
-                }
-                return prop;
-            }
-
-            private SchemaValidationAndTransformConverter GetConverter(MemberInfo member)
-            {
-                var validators = member.GetCustomAttributes<ValidationAttribute>(false);
-                var contentTypeAttributes = member.GetCustomAttributes<DataTypeAttribute>(false);
-                if (contentTypeAttributes.Any() || validators.Any())
-                {
-                    return new SchemaValidationAndTransformConverter(contentTypeAttributes, validators, member.Name);
-                }
-                return null;
-            }
-        }
-
-        private sealed class SchemaValidationAndTransformConverter : JsonConverter
-        {
-            private readonly IEnumerable<ValidationAttribute> _validators;
-            private readonly IEnumerable<DataTypeAttribute> _attributes;
-            private readonly string _fieldName;
-
-            public SchemaValidationAndTransformConverter(IEnumerable<DataTypeAttribute> attributes, IEnumerable<ValidationAttribute> validators, string fieldName)
-            {
-                _attributes = attributes;
-                _validators = validators;
-                _fieldName = fieldName;
-            }
-
-            public override bool CanConvert(Type objectType) => true;
-
-            public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer) => new NotSupportedException();
-
-            public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
-            {
-                var range = ToRange((IJsonLineInfo)reader);
-                var value = serializer.Deserialize(reader, objectType);
-                if (value is null)
-                {
-                    return null;
-                }
-
-                foreach (var validator in _validators)
-                {
-                    try
-                    {
-                        validator.Validate(value, new ValidationContext(value) { DisplayName = _fieldName });
-                    }
-                    catch (Exception e)
-                    {
-                        t_status.Value.Peek().Errors.Add(Errors.ViolateSchema(range, e.Message, reader.Path));
-                    }
-                }
-
-                var transform = t_status.Value.Peek().Transform;
-                return transform != null ? transform(_attributes, value, reader.Path) : value;
-            }
-        }
-
-        private sealed class Status
+        internal class Status
         {
             public List<Error> Errors { get; set; }
